fix order encoding with ethers
