+++ package.json
@@ -11,7 +11,7 @@
     "@nomiclabs/buidler-ethers": "^2.0.0",
     "@nomiclabs/buidler-waffle": "^2.1.0",
     "@openzeppelin/contracts": "^3.2.0",
-    "@types/chai": "^4.2.3",
+    "@types/chai": "^4.2.12",
     "@types/debug": "^4.1.5",
     "@types/mocha": "^8.0.3",
     "@typescript-eslint/eslint-plugin": "^4.3.0",
@@ -25,7 +25,7 @@
     "eslint-plugin-no-only-tests": "^2.4.0",
     "eslint-plugin-prettier": "^3.1.4",
     "ethereum-waffle": "^3.1.1",
-    "ethers": "5.0.14",
+    "ethers": "^5.0.14",
     "prettier": "^2.1.2",
     "ts-node": "^9.0.0",
     "typescript": "^4.0.3"
