--- test/PreAMMBatcher.spec.ts
@@ -1,235 +1,0 @@
-import { ethers, waffle } from "@nomiclabs/buidler";
-import ERC20 from "@openzeppelin/contracts/build/contracts/IERC20.json";
-import { expect } from "chai";
-import { Contract } from "ethers";
-
-import UniswapV2Factory from "../node_modules/@uniswap/v2-core/build/UniswapV2Factory.json";
-import UniswapV2Pair from "../node_modules/@uniswap/v2-core/build/UniswapV2Pair.json";
-import { Order, DOMAIN_SEPARATOR } from "../src/js/orders.spec";
-
-import { baseTestInput } from "./resources/testExamples";
-
-describe("PreAMMBatcher", () => {
-  const [
-    walletDeployer,
-    walletTrader1,
-    walletTrader2,
-  ] = waffle.provider.getWallets();
-  let batcher: Contract;
-  let token0: Contract;
-  let token1: Contract;
-  let uniswapPair: Contract;
-  let uniswapFactory: Contract;
-  let uniswapPairAddress: string;
-  const initialUniswapFundingOfToken0 = ethers.utils.parseEther("10");
-  const initialUniswapFundingOfToken1 = ethers.utils.parseEther("10");
-
-  const mockRelevantTokenDataForBatchTrade = async (
-    sellOrder1: Order,
-    sellOrder2: Order,
-  ): Promise<void> => {
-    // todo: mock actual right information
-    // mock un-relevant data to make batchTrade call pass
-    await token0.mock.transferFrom.returns(true);
-    await token1.mock.transferFrom.returns(true);
-    await token0.mock.transfer.returns(true);
-    await token1.mock.transfer.returns(true);
-    await token0.mock.allowance
-      .withArgs(sellOrder1.wallet.address, batcher.address)
-      .returns(sellOrder1.sellAmount.toString());
-    await token1.mock.allowance
-      .withArgs(sellOrder2.wallet.address, batcher.address)
-      .returns(sellOrder2.sellAmount.toString());
-    await token0.mock.balanceOf
-      .withArgs(sellOrder1.wallet.address)
-      .returns(sellOrder1.sellAmount.toString());
-    await token1.mock.balanceOf
-      .withArgs(sellOrder2.wallet.address)
-      .returns(sellOrder2.sellAmount.toString());
-    await token0.mock.balanceOf.withArgs(batcher.address).returns("1");
-    await token1.mock.balanceOf.withArgs(batcher.address).returns("1");
-    await token0.mock.balanceOf
-      .withArgs(uniswapPair.address)
-      .returns(
-        initialUniswapFundingOfToken0.add(sellOrder1.sellAmount).toString(),
-      );
-    await token1.mock.balanceOf
-      .withArgs(uniswapPair.address)
-      .returns(initialUniswapFundingOfToken1);
-  };
-
-  beforeEach(async () => {
-    const PreAMMBatcher = await ethers.getContractFactory("PreAMMBatcher");
-
-    // deploy all relevant contracts and setup the uniswap pool
-    token0 = await waffle.deployMockContract(walletDeployer, ERC20.abi);
-    token1 = await waffle.deployMockContract(walletDeployer, ERC20.abi);
-    uniswapFactory = await waffle.deployContract(
-      walletDeployer,
-      UniswapV2Factory,
-      [walletDeployer.address],
-    );
-    await uniswapFactory.createPair(token0.address, token1.address, {
-      gasLimit: 6000000,
-    });
-    uniswapPairAddress = await uniswapFactory.getPair(
-      token0.address,
-      token1.address,
-    );
-    uniswapPair = await waffle.deployContract(walletDeployer, UniswapV2Pair);
-    uniswapPair = await uniswapPair.attach(uniswapPairAddress);
-    batcher = await PreAMMBatcher.deploy(uniswapFactory.address);
-
-    await token0.mock.balanceOf
-      .withArgs(uniswapPair.address)
-      .returns(initialUniswapFundingOfToken0);
-    await token1.mock.balanceOf
-      .withArgs(uniswapPair.address)
-      .returns(initialUniswapFundingOfToken1);
-
-    await uniswapPair.mint(walletDeployer.address, { gasLimit: 500000 });
-    expect((await uniswapPair.getReserves())[0]).to.equal(
-      initialUniswapFundingOfToken0,
-    );
-  });
-
-  it("DOMAIN_SEPARATOR is correct", async () => {
-    expect(await batcher.DOMAIN_SEPARATOR()).to.equal(DOMAIN_SEPARATOR);
-  });
-
-  it("orderChecks runs through smoothly", async () => {
-    const testCaseInput = baseTestInput(
-      token0,
-      token1,
-      [walletTrader1],
-      [walletTrader2],
-    );
-
-    await batcher.orderChecks(
-      [testCaseInput.sellOrdersToken0[0].getSmartContractOrder()],
-      [testCaseInput.sellOrdersToken1[0].getSmartContractOrder()],
-      { gasLimit: 6000000 },
-    );
-  });
-
-  it("orderChecks detects non matching orders", async () => {
-    const testCaseInput = baseTestInput(
-      token0,
-      token1,
-      [walletTrader1],
-      [walletTrader2],
-    );
-    testCaseInput.sellOrdersToken1[0].sellToken = token0;
-
-    await expect(
-      batcher.orderChecks(
-        [testCaseInput.sellOrdersToken0[0].getSmartContractOrder()],
-        [testCaseInput.sellOrdersToken1[0].getSmartContractOrder()],
-        { gasLimit: 6000000 },
-      ),
-    ).to.revertedWith("sellOrderToken1 are not compatible in sellToken");
-  });
-
-  it("receiveTradeAmounts reverts if transferFrom fails for token0", async () => {
-    const testCaseInput = baseTestInput(
-      token0,
-      token1,
-      [walletTrader1],
-      [walletTrader2],
-    );
-
-    await mockRelevantTokenDataForBatchTrade(
-      testCaseInput.sellOrdersToken0[0],
-      testCaseInput.sellOrdersToken1[0],
-    );
-    await token0.mock.transferFrom.returns(false);
-    await expect(
-      batcher.batchTrade(
-        testCaseInput.sellOrdersToken0[0].encode(),
-        testCaseInput.sellOrdersToken1[0].encode(),
-        { gasLimit: 6000000 },
-      ),
-    ).to.be.revertedWith("unsuccessful transferFrom for order");
-  });
-
-  it("receiveTradeAmounts transferFrom the right amount of token0", async () => {
-    const testCaseInput = baseTestInput(
-      token0,
-      token1,
-      [walletTrader1],
-      [walletTrader2],
-    );
-
-    await mockRelevantTokenDataForBatchTrade(
-      testCaseInput.sellOrdersToken0[0],
-      testCaseInput.sellOrdersToken1[0],
-    );
-
-    await token0.mock.transferFrom.reverts();
-    await token0.mock.transferFrom
-      .withArgs(
-        walletTrader1.address,
-        batcher.address,
-        testCaseInput.sellOrdersToken0[0].sellAmount,
-      )
-      .returns(true);
-    await expect(
-      batcher.batchTrade(
-        testCaseInput.sellOrdersToken0[0].encode(),
-        testCaseInput.sellOrdersToken1[0].encode(),
-        { gasLimit: 6000000 },
-      ),
-    ).to.not.be.reverted;
-  });
-  it("receiveTradeAmounts reverts if transferFrom fails for token1", async () => {
-    const testCaseInput = baseTestInput(
-      token0,
-      token1,
-      [walletTrader1],
-      [walletTrader2],
-    );
-
-    await mockRelevantTokenDataForBatchTrade(
-      testCaseInput.sellOrdersToken0[0],
-      testCaseInput.sellOrdersToken1[0],
-    );
-
-    await token0.mock.transferFrom.returns(true);
-    await token1.mock.transferFrom.returns(false);
-    await expect(
-      batcher.batchTrade(
-        testCaseInput.sellOrdersToken0[0].encode(),
-        testCaseInput.sellOrdersToken1[0].encode(),
-        { gasLimit: 6000000 },
-      ),
-    ).to.be.revertedWith("unsuccessful transferFrom for order");
-  });
-  it("receiveTradeAmounts transferFrom the right amount of token1", async () => {
-    const testCaseInput = baseTestInput(
-      token0,
-      token1,
-      [walletTrader1],
-      [walletTrader2],
-    );
-    await mockRelevantTokenDataForBatchTrade(
-      testCaseInput.sellOrdersToken0[0],
-      testCaseInput.sellOrdersToken1[0],
-    );
-
-    await token1.mock.transferFrom.reverts();
-    await token1.mock.transferFrom
-      .withArgs(
-        walletTrader2.address,
-        batcher.address,
-        testCaseInput.sellOrdersToken1[0].sellAmount,
-      )
-      .returns(true);
-    await expect(
-      batcher.batchTrade(
-        testCaseInput.sellOrdersToken0[0].encode(),
-        testCaseInput.sellOrdersToken1[0].encode(),
-        { gasLimit: 6000000 },
-      ),
-    ).to.not.be.reverted;
-  });
-});
